refactor: LocalScanner.java를 삭제하고 ConsoleInputer로 옮긴다
